feat(cli): add the agent subcommand namespace
Introduce a new agent command that composes the hook-driven utilities
(triage, wrapup, record) with the three sidecar invocations called by
plugin hook scripts (register-agent, end-agent, inject-env). The parent
command carries a description warning that these commands are intended
for agents and hook scripts rather than direct human use, which the
help formatter renders above the subcommand list.

The sidecar subcommand bodies and their helpers move from the old
hidden internal group into the new agent module. The internal module
is reduced to a transitional shim that re-exports the three sidecar
subcommands so the existing internal group keeps working for one
phase; it is removed once the agent namespace is wired into the bin.
